backend_ml/requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
numpy
python-dotenv
